% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Aligns data. Please note that rows with no alignment day will not be used.}
\usage{
align(
  input,
  out.name = "output_alignment",
  all.markers = T,
  markers = c(),
  write.out = T,
  sheets = F,
  input.format = "csv"
)
}
\arguments{
\item{input}{A file containing all values to be aligned. First column are IDs, second are the alignment days.}

\item{out.name}{The name of the output file. Will be created in .xlsx format. Default "Output_alignment".}

\item{all.markers}{Defines if all markers present in the file should be aligned. If FALSE, please define desired markers by using the markers argument. Default TRUE.}

\item{markers}{A vector containing all desired markers. Upper- and lowercase are not considered. Default is empty.}

\item{write.out}{If TRUE produces an output file in .xlsx format. If FALSE returns the results as data frame. Default TRUE.}

\item{sheets}{If FALSE all markers are copied into one sheet, if TRUE one sheet per marker is created. Default FALSE.}

\item{input.format}{Either "csv", "csv2" or "xlsx" depending on the desired function to read in the input file. Default "csv".}
}
\value{
Either a .xlsx file containing the results, or a data frame depending on the chosen parameters.
}
\description{
Aligns data. Please note that rows with no alignment day will not be used.
}
