% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Bins data.}
\usage{
binning(
  input,
  out.name = "output_binning",
  binning = "scaled",
  x = 2,
  all.markers = TRUE,
  markers = c(),
  sheets = FALSE,
  write.out = TRUE,
  input.format = "csv"
)
}
\arguments{
\item{input}{A file containing all values to be binned. Please make sure that there are no dots in the column names. Depending if the file was aligned prior to binning please change the binning argument.}

\item{out.name}{The name of the output file. Will be created in .xlsx format and contain either s for scale or us for unscaled in the name.}

\item{binning}{Defines if the input data has been aligned beforehand. If so, use 'scaled, if the data has not been aligned, use 'unscaled'. Please note that for a data that already has been binned, all data has to be in one (the first) sheet. Multiple sheets will not be recognized. Default is 'scaled'.}

\item{x}{Number of days to be binned by. Default is 2.}

\item{all.markers}{Defines if all markers present in the file should be aligned. If FALSE, please define desired markers by using the markers argument. Default TRUE.}

\item{markers}{A vector containing all desired markers. Upper- and lowercase are not considered. Default is empty.}

\item{sheets}{If FALSE all markers are copied into one sheet, if TRUE one sheet per marker is created. Default FALSE.}

\item{write.out}{If TRUE produces an outputfile in .xlsx format. If FALSE returns the results as data frame. Default TRUE.}

\item{input.format}{Either "csv", "csv2" or "xlsx" depending on the desired function to read in the input file. Default "csv".}
}
\value{
Either a .xlsx file containing the results, or a data frame depending on the chosen parameters.
}
\description{
Bins data.
}
